CRUD updated and working with postman

+++ controllers/postController.js
@@ -3,7 +3,7 @@ const validationHandler = require("../validations/validationHandler");
 
 exports.index = async (req, res) => { //this will show the posts
     try {
-        const posts = await Post.find().sort({ createdAt: -1 }); //newer will come first
+        const posts = await Post.find().populate("user").sort({ createdAt: -1 }); //newer will come first
         res.send(posts)
     }catch(err){
         next(err);
@@ -14,7 +14,7 @@ exports.show = async (req, res) => { //shows a post based in one id
     try{
       const post = await Post.findOne({
           _id: req.params.id
-       });
+       }).populate("user");
 
        res.send(post);
       
@@ -30,6 +30,7 @@ exports.store = async (req, res, next) => { //stores the post in our DB
         let post = new Post();
         post.description = req.body.description;
         post.image = req.file.filename;
+        post.user = req.user;
         post = await post.save();
 
         res.send(post);
@@ -43,6 +44,11 @@ exports.update = async (req, res, next) => { //changes values from a post, commo
         validationHandler(req);
 
         let post = await Post.findById(req.params.id);
+        if (!post || post.user != req.user.id) {
+            const error = new Error("Wrong request");
+            error.statusCode = 400;
+            throw error;
+        }
         post.description = req.body.description;
         post = await post.save();
 
@@ -56,6 +62,11 @@ exports.delete = async (req, res, next) => { //deletes a post, commonly if is te
     try{
 
         let post = await Post.findById(req.params.id);
+        if (!post || post.user != req.user.id) {
+            const error = new Error("Wrong request");
+            error.statusCode = 400;
+            throw error;
+        }
         await post.delete();
 
         res.send({ message: "success" });
